% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{CalcPsiRMST}
\alias{CalcPsiRMST}
\title{Integrate Kaplan Meier}
\usage{
CalcPsiRMST(status, time, tau)
}
\arguments{
\item{status}{Status, coded as 0 for censoring, 1 for death.}

\item{time}{Observation time.}

\item{tau}{Truncation time.}
}
\description{
Integrate the Kaplan-Meier curve between two values.
}
