% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenPerturb}
\alias{GenPerturb}
\title{Generate Perturbations}
\usage{
GenPerturb(psi, n_boot)
}
\arguments{
\item{psi}{Per-subject influence values.}

\item{n_boot}{Number of perturbations.}
}
\value{
Numeric vector.
}
\description{
Generate Perturbations
}
\section{Notes}{

The random seed should be set in R prior to calling this function.
}

