% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilities.R
\name{MaxEventTime}
\alias{MaxEventTime}
\title{Maximum Event Time}
\usage{
MaxEventTime(statuses, times)
}
\arguments{
\item{statuses}{Array with 1 or more status variables.}

\item{times}{Array with 1 or more event-time variables.}
}
\value{
Numeric maximum event time.
}
\description{
Find the maximum time with a status == 1.
}
