% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MRMST-package.R
\docType{package}
\name{MRMST-package}
\alias{MRMST}
\alias{MRMST-package}
\title{MRMST: Multiple Restricted Mean Survival Time}
\description{
Inference on the multiple restricted mean survival times.
}
\author{
\strong{Maintainer}: Zachary McCaw \email{zmccaw@alumni.harvard.edu} (\href{https://orcid.org/0000-0002-2006-9828}{ORCID})

}
\keyword{internal}
