% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{OneSample-class}
\alias{OneSample-class}
\title{One Sample}
\description{
One Sample
}
\section{Slots}{

\describe{
\item{\code{AUC}}{Area under the curve.}

\item{\code{K}}{Number of outcomes.}

\item{\code{SE}}{Standard error.}

\item{\code{Table}}{Table.}

\item{\code{Tau}}{Truncation time.}
}}

