% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/class.R
\docType{class}
\name{TwoSample-class}
\alias{TwoSample-class}
\title{Two Sample}
\description{
Two Sample
}
\section{Slots}{

\describe{
\item{\code{Arm0}}{Results for the reference arm.}

\item{\code{Arm1}}{Results for the treatment arm.}

\item{\code{Contrast}}{Contrasts.}
}}

