% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/two_sample.R
\name{TwoSample}
\alias{TwoSample}
\title{Two Sample Calculation}
\usage{
TwoSample(
  arm,
  statuses,
  times,
  alpha = 0.05,
  extend = FALSE,
  n_boot = 2000,
  tau = NULL,
  weights = NULL
)
}
\arguments{
\item{arm}{Treatment arm, coded as 0 for reference, 1 for treatment.}

\item{statuses}{Array with 1 or more status variables.}

\item{times}{Array with 1 or more event-time variables.}

\item{alpha}{Type I error.}

\item{extend}{Extend the Kaplan-Meier curves if \code{tau} is set beyond the
maximum observation time?}

\item{n_boot}{Number of perturbations for variance calculation.}

\item{tau}{Truncation time.}

\item{weights}{Optional weight vector. Defaults to equal weights.}
}
\description{
Two Sample Calculation
}
