% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{GenPerturb}
\alias{GenPerturb}
\title{Generate Perturbations

Generates realizations of \eqn{\frac{1}{n}\sum_{i=1}^{n}\psi_{i}w_{i}},
where \eqn{\psi_{i}} is the influence function for the ith subject and the
\eqn{w_{i}} are IID random weights.}
\usage{
GenPerturb(psi, n_boot)
}
\arguments{
\item{psi}{Per-subject influence values.}

\item{n_boot}{Number of perturbations.}
}
\value{
Numeric vector.
}
\description{
Generate Perturbations

Generates realizations of \eqn{\frac{1}{n}\sum_{i=1}^{n}\psi_{i}w_{i}},
where \eqn{\psi_{i}} is the influence function for the ith subject and the
\eqn{w_{i}} are IID random weights.
}
\section{Notes}{

The random seed should be set in R prior to calling this function.
}

